+++ src/geometryRTheta/time_solver/bsl_predcorr_second_order_implicit.hpp
@@ -391,7 +391,7 @@ public:
 
 
     /**
-     * @brief The implicit loop which calculates the feet of the charateristics.
+     * @brief The implicit loop which calculates the feet of the characteristicss.
      *
      * This function should be private but cannot be as it contains Kokkos lambda
      * functions.